server/requirements.txt
alembic==1.11.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
asyncpg==0.28.0
Authlib==1.2.1
backcall==0.2.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
expects==0.9.0
fastapi==0.103.2
flake8==6.0.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
iniconfig==2.0.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.0.0
pluggy==1.3.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.0.2
pydantic_core==2.14.5
pyflakes==3.0.1
Pygments==2.17.2
PyJWT==2.7.0
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
stack-data==0.6.3
starlette==0.27.0
traitlets==5.14.0
types-passlib==1.7.7.12
types-requests==2.31.0.7
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.22.0
wcwidth==0.2.12
python-semantic-release==9.8.6
fastapi-mail==1.4.1
gunicorn==23.0.0
flagsmith==3.8.0